Move styles to layout folder for re-usablility
