+++ src/App.js
@@ -1,16 +1,18 @@
 import React, { useState } from "react";
 import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
 import { ThemeProvider } from "styled-components";
-import Home from "pages/Home";
-import Explore from "pages/Explore";
-import NavBar from "components/Header/NavBar";
-import User from "pages/User";
-import SearchPhotoResult from "pages/SearchPhotoResult";
-import SearchCollectionResult from "pages/SearchCollectionResult";
-import Topic from "pages/Topic";
-import FavoriteImage from "pages/FavoriteImage";
+import { NavBar } from "components";
 import { GlobalStyle, lightTheme, darkTheme, AppContainer } from "App.styles";
-import Collection from "pages/Collection";
+import {
+  Home,
+  Explore,
+  SearchPhotoResult,
+  SearchCollectionResult,
+  Topic,
+  User,
+  FavoriteImage,
+  Collection,
+} from "pages";
 
 export default function App() {
   const [theme, setTheme] = useState(true);
@@ -32,10 +34,7 @@ export default function App() {
               path="/search/collections/:searchTerm"
               component={SearchCollectionResult}
             />
-            <Route
-              path="/topic/:searchTerm"
-              component={Topic}
-            />
+            <Route path="/topic/:searchTerm" component={Topic} />
             <Route
               path="/collection/:id&total_photos=:total_photos&user=:user&title=:title"
               component={Collection}
