+++ src/components/UI/Layout/SearchPageInfoLayout.styles.js
@@ -9,29 +9,6 @@ export const DisplayArea = styled.div`
   background-color: #f9fafb;
 `;
 
-export const ImageColumn = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  margin: 15px;
-`;
-
-export const ImageArea = styled.div`
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  padding-top: 20px;
-  padding-left: 14px;
-  padding-right: 14px;
-  width: 928px;
-`;
-
-export const ImageContainer = styled.div`
-  display: flex;
-  justify-content: center;
-  flex-direction: row;
-`;
-
 export const PhotosAndSelectionsContainer = styled.div`
   display: flex;
   justify-content: center;
@@ -45,6 +22,7 @@ export const StyledLink = styled(Link)`
   color: #7B7B7B;
   opacity: 0.39;
   text-decoration: none;
+  
 `;
 
 export const UnderScoredLink = styled(Link)`
