--- src/pages/Explore/Explore.styles.js
@@ -1,33 +1,0 @@
-import styled from "styled-components";
-
-export const DisplayArea = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  justify-self: center;
-  background-color: #f9fafb;
-`;
-
-export const ImageContainer = styled.div`
-  display: flex;
-  justify-content: center;
-  flex-direction: row;
-`;
-
-export const ImageArea = styled.div`
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  padding-top: 20px;
-  padding-left: 14px;
-  padding-right: 14px;
-  width: 928px;
-`;
-
-export const ImageColumn = styled.div`
-  display: flex;
-  flex-direction: column;
-  flex-basis: 100%;
-  align-items: center;
-  margin: 15px;
-`;
