+++ src/pages/User/User.styles.js
@@ -1,33 +1,5 @@
 import styled from "styled-components";
 
-export const DisplayArea = styled.div`
-  display: flex;
-  flex-direction: column;
-`;
-
-export const ImageColumn = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  margin: 15px;
-`;
-
-export const ImageArea = styled.div`
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  padding-top: 20px;
-  padding-left: 14px;
-  padding-right: 14px;
-  width: 928px;
-`;
-
-export const ImageContainer = styled.div`
-  display: flex;
-  justify-content: center;
-  flex-direction: row;
-`;
-
 export const UserInfoContainer = styled.div`
   display: flex;
   align-items: center;
