+++ src/pages/index.js
@@ -1,0 +1,8 @@
+export { default as Home } from "./Home";
+export { default as Explore } from "./Explore";
+export { default as SearchPhotoResult } from "./SearchPhotoResult";
+export { default as SearchCollectionResult } from "./SearchCollectionResult";
+export { default as Topic } from "./Topic";
+export { default as FavoriteImage } from "./FavoriteImage";
+export { default as Collection } from "./Collection";
+export { default as User } from "./User";
